backend/requirements-PRODUCTION.txt
psycopg2==2.6.1
django-storages==1.1.8
boto==2.38.0
